+++ include/scene/objects/quad.hpp
@@ -33,21 +33,21 @@ class quad : public object {
         
         // Constructor from four points
         quad(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2, const rt::vector& p3,
-            const size_t material_index, const std::optional<texture_info>& info);
+            const unsigned int material_index, const std::optional<texture_info>& info);
 
         // Constructor from four points with vertex normals
         quad(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2, const rt::vector& p3,
             const rt::vector& vn0, const rt::vector& vn1, const rt::vector& vn2, const rt::vector& vn3,
-            const size_t material_index, const std::optional<texture_info>& info);
+            const unsigned int material_index, const std::optional<texture_info>& info);
 
         // Constructor from four points with normal mapping enabled
         quad(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2, const rt::vector& p3,
-            const size_t material_index, const std::optional<texture_info>& info, const bool normal_mapping);
+            const unsigned int material_index, const std::optional<texture_info>& info, const bool normal_mapping);
 
         // Constructor from four points with vertex normals and normal mapping enabled
         quad(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2, const rt::vector& p3,
             const rt::vector& vn0, const rt::vector& vn1, const rt::vector& vn2, const rt::vector& vn3,
-            const size_t material_index, const std::optional<texture_info>& info, const bool normal_mapping);
+            const unsigned int material_index, const std::optional<texture_info>& info, const bool normal_mapping);
 
         /* Returns barycenter of the quad */
         rt::vector get_barycenter() const;
