+++ src/file_readers/mtl_parser.cpp
@@ -24,7 +24,7 @@
 bool parse_mtl_file(const char* file_name, const std::string& path,
     std::vector<wrapper<material>>& material_wrapper_set,
     std::vector<wrapper<texture>>& texture_wrapper_set,
-    std::map<size_t, size_t>& mt_assoc, const real gamma) {
+    std::map<unsigned int, unsigned int>& mt_assoc, const real gamma) {
 
     FILE* file = fopen((path + std::string(file_name)).data(), "r");
 
@@ -141,7 +141,7 @@ bool parse_mtl_file(const char* file_name, const std::string& path,
                     material_wrapper_set.emplace_back(std::move(m), m_name);
                 }
                 
-                const size_t m_i = material_wrapper_set[material_wrapper_set.size()-1].index;
+                const unsigned int m_i = material_wrapper_set[material_wrapper_set.size()-1].index;
 
                 /* Test for associated texture */
                 char tfile_name[513];
@@ -153,7 +153,7 @@ bool parse_mtl_file(const char* file_name, const std::string& path,
                     bool parsing_successful;
                     texture txt((path + std::string(tfile_name)).data(), parsing_successful);
                     texture_wrapper_set.emplace_back(std::move(txt));
-                    const size_t t_i = texture_wrapper_set[texture_wrapper_set.size()-1].index;
+                    const unsigned int t_i = texture_wrapper_set[texture_wrapper_set.size()-1].index;
 
                     if (parsing_successful) {
                         printf("\rmtl_parser: %s texture loaded\n", tfile_name);
