+++ src/file_readers/obj_parser.cpp
@@ -47,10 +47,10 @@ void add_triangle(const std::vector<rt::vector>& vertex_set, const std::vector<r
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_triangles,
    const rt::vector& shift, const real scale,
-   const size_t v1,  const size_t v2,  const size_t v3,
-   const size_t vt1, const size_t vt2, const size_t vt3,
-   const size_t vn1, const size_t vn2, const size_t vn3,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int v1,  const unsigned int v2,  const unsigned int v3,
+   const unsigned int vt1, const unsigned int vt2, const unsigned int vt3,
+   const unsigned int vn1, const unsigned int vn2, const unsigned int vn3,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    const bool apply_texture) {
 
    const std::optional<texture_info> info =
@@ -88,10 +88,10 @@ void add_triangle_subdiv(const std::vector<rt::vector>& vertex_set, const std::v
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_triangles,
    const rt::vector& shift, const real scale,
-   const size_t vj,  const size_t vi,  const rt::vector& final_v,
-   const size_t vtj, const size_t vti, const rt::vector& final_vt,
-   const size_t vnj, const size_t vni, const rt::vector& final_vn,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int vj,  const unsigned int vi,  const rt::vector& final_v,
+   const unsigned int vtj, const unsigned int vti, const rt::vector& final_vt,
+   const unsigned int vnj, const unsigned int vni, const rt::vector& final_vn,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    const bool apply_texture) {
 
    const std::optional<texture_info> info =
@@ -129,10 +129,10 @@ void add_quad(const std::vector<rt::vector>& vertex_set, const std::vector<rt::v
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_quads,
    const rt::vector& shift, const real scale,
-   const size_t v1,  const size_t v2,  const size_t v3,  const size_t v4,
-   const size_t vt1, const size_t vt2, const size_t vt3, const size_t vt4,
-   const size_t vn1, const size_t vn2, const size_t vn3, const size_t vn4,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int v1,  const unsigned int v2,  const unsigned int v3,  const unsigned int v4,
+   const unsigned int vt1, const unsigned int vt2, const unsigned int vt3, const unsigned int vt4,
+   const unsigned int vn1, const unsigned int vn2, const unsigned int vn3, const unsigned int vn4,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    const bool apply_texture) {
 
    const std::optional<texture_info> info =
@@ -172,9 +172,9 @@ void add_triangle_no_normal(const std::vector<rt::vector>& vertex_set, const std
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_triangles,
    const rt::vector& shift, const real scale,
-   const size_t v1,  const size_t v2,  const size_t v3,
-   const size_t vt1, const size_t vt2, const size_t vt3,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int v1,  const unsigned int v2,  const unsigned int v3,
+   const unsigned int vt1, const unsigned int vt2, const unsigned int vt3,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    const bool apply_texture) {
 
    const std::optional<texture_info> info =
@@ -210,9 +210,9 @@ void add_triangle_subdiv_no_normal(const std::vector<rt::vector>& vertex_set, co
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_triangles,
    const rt::vector& shift, const real scale,
-   const size_t vj,  const size_t vi,  const rt::vector& final_v,
-   const size_t vtj, const size_t vti, const rt::vector& final_vt,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int vj,  const unsigned int vi,  const rt::vector& final_v,
+   const unsigned int vtj, const unsigned int vti, const rt::vector& final_vt,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    const bool apply_texture) {
 
    const std::optional<texture_info> info =
@@ -248,9 +248,9 @@ void add_quad_no_normal(const std::vector<rt::vector>& vertex_set, const std::ve
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_quads,
    const rt::vector& shift, const real scale,
-   const size_t v1,  const size_t v2,  const size_t v3,  const size_t v4,
-   const size_t vt1, const size_t vt2, const size_t vt3, const size_t vt4,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int v1,  const unsigned int v2,  const unsigned int v3,  const unsigned int v4,
+   const unsigned int vt1, const unsigned int vt2, const unsigned int vt3, const unsigned int vt4,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    const bool apply_texture) {
 
    const std::optional<texture_info> info =
@@ -291,15 +291,15 @@ void add_subdivided_polygon(FILE* file,
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_triangles,
    const rt::vector& shift, const real scale,
-   const size_t v1, const size_t v2, const size_t v3, const size_t v4, const size_t v5,
-   const size_t vt1, const size_t vt2, const size_t vt3, const size_t vt4, const size_t vt5,
-   const size_t vn1, const size_t vn2, const size_t vn3, const size_t vn4, const size_t vn5,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int v1, const unsigned int v2, const unsigned int v3, const unsigned int v4, const unsigned int v5,
+   const unsigned int vt1, const unsigned int vt2, const unsigned int vt3, const unsigned int vt4, const unsigned int vt5,
+   const unsigned int vn1, const unsigned int vn2, const unsigned int vn3, const unsigned int vn4, const unsigned int vn5,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    bool apply_texture) {
    
-   std::stack<size_t> v_stack;
-   std::stack<size_t> vt_stack;
-   std::stack<size_t> vn_stack;
+   std::stack<unsigned int> v_stack;
+   std::stack<unsigned int> vt_stack;
+   std::stack<unsigned int> vn_stack;
    v_stack.push(v1);   v_stack.push(v2);   v_stack.push(v3);   v_stack.push(v4);   v_stack.push(v5);
    vt_stack.push(vt1); vt_stack.push(vt2); vt_stack.push(vt3); vt_stack.push(vt4); vt_stack.push(vt5);
    vn_stack.push(vn1); vn_stack.push(vn2); vn_stack.push(vn3); vn_stack.push(vn4); vn_stack.push(vn5);
@@ -355,26 +355,26 @@ void add_subdivided_polygon(FILE* file,
    final_vn = final_vn / cpt;
 
    // Keeping the last vertex in memory to form a triangle with the first vertex
-   const size_t last_v  = v_stack.top();
-   size_t last_vt = 0;
+   const unsigned int last_v  = v_stack.top();
+   unsigned int last_vt = 0;
    if (apply_texture) { last_vt = vt_stack.top(); }
-   const size_t last_vn = vn_stack.top();
+   const unsigned int last_vn = vn_stack.top();
 
    // Adding the new triangles having the new central vertex as a common vertex
-   for (size_t i = 0; i < cpt - 1; i++) {
-      const size_t vi = v_stack.top();
-      size_t vti = 0;
+   for (unsigned int i = 0; i < cpt - 1; i++) {
+      const unsigned int vi = v_stack.top();
+      unsigned int vti = 0;
       if (apply_texture) { vti = vt_stack.top(); }
-      const size_t vni = vn_stack.top();
+      const unsigned int vni = vn_stack.top();
 
       v_stack.pop();
       if (apply_texture) { vt_stack.pop(); }
       vn_stack.pop();
 
-      const size_t vj = v_stack.top();
-      size_t vtj = 0;
+      const unsigned int vj = v_stack.top();
+      unsigned int vtj = 0;
       if (apply_texture) { vtj = vt_stack.top(); }
-      const size_t vnj = vn_stack.top();
+      const unsigned int vnj = vn_stack.top();
 
       add_triangle_subdiv(vertex_set, uv_coord_set, normal_set, obj_set, content, bounding_enabled,
          number_of_polygons, number_of_triangles,
@@ -402,13 +402,13 @@ void add_subdivided_polygon_no_normal(FILE* file,
    std::vector<const object*>& obj_set, std::vector<const object*>& content, const bool bounding_enabled,
    unsigned int& number_of_polygons, unsigned int& number_of_triangles,
    const rt::vector& shift, const real scale,
-   const size_t v1,  const size_t v2,  const size_t v3,  const size_t v4,  const size_t v5,
-   const size_t vt1, const size_t vt2, const size_t vt3, const size_t vt4, const size_t vt5,
-   const size_t current_texture_index, const size_t current_material_index,
+   const unsigned int v1,  const unsigned int v2,  const unsigned int v3,  const unsigned int v4,  const unsigned int v5,
+   const unsigned int vt1, const unsigned int vt2, const unsigned int vt3, const unsigned int vt4, const unsigned int vt5,
+   const unsigned int current_texture_index, const unsigned int current_material_index,
    bool apply_texture) {
    
-   std::stack<size_t> v_stack;
-   std::stack<size_t> vt_stack;
+   std::stack<unsigned int> v_stack;
+   std::stack<unsigned int> vt_stack;
    v_stack.push(v1);   v_stack.push(v2);   v_stack.push(v3);   v_stack.push(v4);   v_stack.push(v5);
    vt_stack.push(vt1); vt_stack.push(vt2); vt_stack.push(vt3); vt_stack.push(vt4); vt_stack.push(vt5);
 
@@ -453,21 +453,21 @@ void add_subdivided_polygon_no_normal(FILE* file,
    if (apply_texture) { final_vt = final_vt / cpt; }
 
    // Keeping the last vertex in memory to form a triangle with the first vertex
-   const size_t last_v  = v_stack.top();
-   size_t last_vt = 0;
+   const unsigned int last_v  = v_stack.top();
+   unsigned int last_vt = 0;
    if (apply_texture) { last_vt = vt_stack.top(); }
 
    // Adding the new triangles having the new central vertex as a common vertex
-   for (size_t i = 0; i < cpt - 1; i++) {
-      const size_t vi = v_stack.top();
-      size_t vti = 0;
+   for (unsigned int i = 0; i < cpt - 1; i++) {
+      const unsigned int vi = v_stack.top();
+      unsigned int vti = 0;
       if (apply_texture) { vti = vt_stack.top(); }
 
       v_stack.pop();
       if (apply_texture) { vt_stack.pop(); }
 
-      const size_t vj = v_stack.top();
-      size_t vtj = 0;
+      const unsigned int vj = v_stack.top();
+      unsigned int vtj = 0;
       if (apply_texture) { vtj = vt_stack.top(); }
 
       add_triangle_subdiv_no_normal(vertex_set, uv_coord_set, obj_set, content, bounding_enabled,
@@ -502,7 +502,7 @@ void add_subdivided_polygon_no_normal(FILE* file,
    Returns true if the operation was successful
 */
 
-bool parse_obj_file(const char* file_name, const std::optional<size_t> default_texture_index,
+bool parse_obj_file(const char* file_name, const std::optional<unsigned int> default_texture_index,
    std::vector<const object*>& obj_set,
    std::vector<wrapper<material>>& material_wrapper_set,
    std::vector<wrapper<texture>>& texture_wrapper_set,
@@ -522,7 +522,7 @@ bool parse_obj_file(const char* file_name, const std::optional<size_t> default_t
 
    /* Extraction of the path to the .obj file, to be appended to relative paths of mtl and texture files */
    const string file_name_string = string(file_name);
-   const size_t last_slash = file_name_string.find_last_of("/\\");
+   const unsigned int last_slash = file_name_string.find_last_of("/\\");
    const string path = file_name_string.substr(0, last_slash + 1);
 
    /* Storage */
@@ -538,10 +538,10 @@ bool parse_obj_file(const char* file_name, const std::optional<size_t> default_t
    std::vector<rt::vector> normal_set = { rt::vector() };
 
    /* Material -> texture association table */
-   map<size_t, size_t> mt_assoc;
+   map<unsigned int, unsigned int> mt_assoc;
 
-   size_t current_material_index = 0;
-   size_t current_texture_index = default_texture_index.value_or((size_t) (-1));
+   unsigned int current_material_index = 0;
+   unsigned int current_texture_index = default_texture_index.value_or((unsigned int) (-1));
 
    const bool default_texture_provided = default_texture_index.has_value();
    bool apply_texture = default_texture_provided;
@@ -679,7 +679,7 @@ bool parse_obj_file(const char* file_name, const std::optional<size_t> default_t
             }
 
             /* Looking up the material name in the vector of already declared material names */
-            std::optional<size_t> vindex = std::nullopt;
+            std::optional<unsigned int> vindex = std::nullopt;
             for (wrapper<material> const& mat_wrap : material_wrapper_set) {
                if (mat_wrap.name.has_value() && mat_wrap.name.value().compare(m_name) == 0) {
                   vindex = mat_wrap.index;
@@ -698,7 +698,7 @@ bool parse_obj_file(const char* file_name, const std::optional<size_t> default_t
                   apply_texture = true;
                }
                else {
-                  current_texture_index = default_texture_index.value_or((size_t) (-1));
+                  current_texture_index = default_texture_index.value_or((unsigned int) (-1));
                   apply_texture = default_texture_provided;
                }
             }
