+++ src/file_readers/raw_data.cpp
@@ -29,8 +29,8 @@ bool export_raw(const char* file_name, const unsigned int number_of_rays, std::v
 
     try {
     
-        const size_t width = matrix.size();
-        const size_t height = matrix[0].size();
+        const unsigned int width = matrix.size();
+        const unsigned int height = matrix[0].size();
 
         const int ret0 = fprintf(file, "width:%u height:%u number_of_rays:%u\n",
             (unsigned int) width, (unsigned int) height, number_of_rays);
@@ -40,12 +40,12 @@ bool export_raw(const char* file_name, const unsigned int number_of_rays, std::v
             throw std::runtime_error("");
         }
 
-        for(size_t j = 0; j < height; j++) {
-            for(size_t i = 0; i < width; i++) {
+        for(unsigned int j = 0; j < height; j++) {
+            for(unsigned int i = 0; i < width; i++) {
                 rt::color c = matrix[i][j];
                 const int ret = fprintf(file, "%lf %lf %lf\n", c.get_red(), c.get_green(), c.get_blue());
                 if (ret < 0) {
-                    printf("Writing error at color line %zu of %s\n", width*j + i, file_name);
+                    printf("Writing error at color line %u of %s\n", width*j + i, file_name);
                     throw std::runtime_error("");
                 }
             }
@@ -85,12 +85,12 @@ std::optional<std::vector<std::vector<rt::color>>> read_raw(const char* file_nam
 
         std::vector<std::vector<rt::color>> matrix(width, std::vector<rt::color>(height));
 
-        for(size_t j = 0; j < height; j++) {
-            for(size_t i = 0; i < width; i++) {
+        for(unsigned int j = 0; j < height; j++) {
+            for(unsigned int i = 0; i < width; i++) {
                 double r, g, b;
                 const int ret = fscanf(file, "%lf %lf %lf\n", &r, &g, &b);
                 if (ret < 0) {
-                    printf("Reading error at color line %zu of file %s\n", width*j + i, file_name);
+                    printf("Reading error at color line %u of file %s\n", width*j + i, file_name);
                     throw std::runtime_error("");
                 }
                 matrix[i][j] = rt::color(r, g, b);
@@ -164,13 +164,13 @@ bool combine_raw(const char* dest_bmp_name, const char* dest_raw_name, const int
 
         total_number_of_rays += number_of_rays_k;
 
-        for(size_t j = 0; j < height; j++) {
-            const size_t widthj = width * j;
-            for(size_t i = 0; i < width; i++) {
+        for(unsigned int j = 0; j < height; j++) {
+            const unsigned int widthj = width * j;
+            for(unsigned int i = 0; i < width; i++) {
                 double r, g, b;
                 const int ret = fscanf(file, "%lf %lf %lf\n", &r, &g, &b);
                 if (ret < 0) {
-                    printf("Reading error at color line %zu of file %s\n", widthj + i, source_file_names[k]);
+                    printf("Reading error at color line %u of file %s\n", widthj + i, source_file_names[k]);
                     fclose(file);
                     return false;
                 }
