+++ src/scene/objects/cylinder.cpp
@@ -12,7 +12,7 @@ cylinder::cylinder()
     : radius(0), length(0) {}
 
 cylinder::cylinder(const rt::vector& origin, const rt::vector& direction,
-    const real radius, const real length, const size_t material_index)
+    const real radius, const real length, const unsigned int material_index)
 
     : object(origin, material_index),
         direction(direction), radius(radius), length(length) {}
